(issue14) Move user_agent to respective app/agents folder
