+++ agents/services/planner_service.py
@@ -1,11 +1,3 @@
-import sys
-import os
-
-sys.path.insert(
-    0,
-    os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")),
-)
-
 from autogen import AssistantAgent, UserProxyAgent
