+++ tests/agents/services/chat_service_test.py
@@ -10,7 +10,7 @@ def test_initiate_chat(mocked_wait, mocked_create):
     mocked_wait.return_value = None
 
     with patch("autogen.AssistantAgent") as MockedAssistant, patch(
-        "agents.user_agent.UserAgent"
+        "app.agents.user_agent.UserAgent"
     ) as MockedProxy:
         service = ChatService(
             "some_config", {}, MockedAssistant.return_value, MockedProxy.return_value
