+++ tests/app/agents/user_agent_test.py
@@ -1,9 +1,4 @@
-import sys
-import os
-
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
-
-from agents.user_agent import UserAgent
+from app.agents.user_agent import UserAgent
 
 import pytest
